Move postAccessRequired tests to the separate file
